Add files via upload
I managed to fix the icon in the create tournament button on the Teamlist page. Created a functioning button that can make games complete before the end date as long as there is a final winner. And added functionality to the anonymous/showall page, it might need some updating css though. Uploading to github now.